requirements.txt
kaggle
firebase-admin
python-dotenv
numpy
pandas
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
youtube-transcript-api
pytrends
schedule
transformers
peft
datasets
wandb
torch
torchvision
torchaudio
